requirements.txt
Django~=3.1.4
channels~=3.0.2
Twisted~=20.3.0
Pillow~=8.0.1
psycopg2~=2.8.6